+++ pySimBlocks/blocks/observers/luenberger.py
@@ -57,10 +57,10 @@ class Luenberger(Block):
         # ------------------------------------------------------------------
         # Store and check matrices
         # ------------------------------------------------------------------
-        self.A = np.asarray(A)
-        self.B = np.asarray(B)
-        self.C = np.asarray(C)
-        self.L = np.asarray(L)
+        self.A = np.asarray(A, dtype=float)
+        self.B = np.asarray(B, dtype=float)
+        self.C = np.asarray(C, dtype=float)
+        self.L = np.asarray(L, dtype=float)
 
         n = self.A.shape[0]
 
@@ -83,9 +83,9 @@ class Luenberger(Block):
         # Initial state x0
         # ------------------------------------------------------------------
         if x0 is None:
-            x0 = np.zeros((n, 1))
+            x0 = np.zeros((n, 1), dtype=float)
         else:
-            x0 = np.asarray(x0).reshape(-1, 1)
+            x0 = np.asarray(x0, dtype=float).reshape(-1, 1)
             if x0.shape != (n, 1):
                 raise ValueError(f"x0 must have shape ({n}, 1).")
