+++ pySimBlocks/blocks/operators/dead_zone.py
@@ -64,7 +64,7 @@ class DeadZone(Block):
 
     # ------------------------------------------------------------------
     def _to_column(self, name, value):
-        arr = np.asarray(value)
+        arr = np.asarray(value, dtype=float)
 
         if arr.ndim == 0:
             return arr.reshape(1, 1)
