+++ pySimBlocks/blocks/operators/discrete_derivator.py
@@ -43,7 +43,7 @@ class DiscreteDerivator(Block):
         self._first_output = True
 
         if initial_output is not None:
-            arr = np.asarray(initial_output)
+            arr = np.asarray(initial_output, dtype=float)
             if arr.ndim == 0:
                 arr = arr.reshape(1,1)
             elif arr.ndim == 1:
