+++ pySimBlocks/blocks/operators/saturation.py
@@ -51,7 +51,7 @@ class Saturation(Block):
 
     # ------------------------------------------------------------------
     def _to_column(self, name, value):
-        arr = np.asarray(value)
+        arr = np.asarray(value, dtype=float)
 
         if arr.ndim == 0:
             return arr.reshape(1, 1)
